Improve rgb_init function and add support for SPI over PIO in main.
The rgb_init function no longer requires the sm parameter, it now internally handles the allocation of an unused state machine. This change simplifies its usage. Main.cpp file was updated to include an SPI interface over PIO and the use of the PIOSPIInterface and LCDDirectGraphics classes. Additionally, a resetRotation method was added to the RotaryEncoder class for flexibility.