Update USB identifiers and keyboard behavior
The USB Vendor and Product Identifiers have been updated in usb_descriptors.cpp. The behavior for key A press on button in main.cpp has been changed to simulate a Shift + A press instead of just A. Also the method to measure the interval in main.cpp has been updated to use time_us_64() function.